KFSCNTRB-1996 Never treat Vendors as Employees when extracting to PDP
The existing code treats a DV payee as an Employee if at all possible, that is
even if the payee is a vendor, if he is also an employee, treat him as an employee.
We would rather treat the payee as a vendor if possible so that we collect things
like the ownership code and the possibility of ACH payments (which aren't possible
for employees apparently).

git-svn-id: https://svn.kuali.org/repos/kfs/branches/release-5-3-1@45377 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7